Update `capture` method of `Facade` class
Prior to this change, captured amount was equal to charged amount. But
this is not always the case - sometimes lower amount must be captured.

* If `amount` passed to `capture`, it will be used.
* If `amount` does not passed, will be used `order.total_incl_tax`. This
needed because in some cases `order.total_incl_tax` is not equal to the
amount pre-authorized during `charging`. E.g. `order.total_incl_tax`
was pre-authorized during order placement. But after this some discount
was applied to this order. So now `order.total_incl_tax` keeps value
that different (lower) from pre-authorized one.